Force "byte encoding" in SMPP UDH
UDH bytes are being prepended to SMPP message parts when we send
long messages. But they're being treated as US-ASCII (7 bit)
encoded, so that breaks when trying to parse it for concatenation
with the part's text.

Forcing the encoding to be ASCII-8BIT (aka, "just bytes") we get
rid of the issue.

Fixes #32